calculadora_financeira/pyproject.toml
[project]
name = "calculadora-financeira"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "fastmcp>=2.12.3",
    "mysql-connector-python>=9.4.0",
]
